+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Renamed `WaitForComponent` to `WaitForExecutor`.
+
 ## [1.0.4a1] - 2021-11-22
 ### Added
 - `prefix_match` option to ComponentClient custom ids to make storing metadata in custom ids